Combine edge state and attempt gate into attention
