+++ cognition/attempt_gate.py
@@ -1,223 +1,11 @@
-"""Bounded self-assessment gate for execution paths.
+"""Compatibility wrapper for attention gate helpers.
 
-Used by EdgeCognitiveState.should_attempt and AikoThink.route / agentic_chat.
-Keeps thresholds conservative: critical work always proceeds.
-
-This module is intentionally outside cognition.memory because it does not read
-or write long-term memory.  It is a pure decision helper over the bounded
-per-user state snapshot: energy/load readiness, uncertainty, recent tool
-outcomes, contradictions, time sensitivity, answer-completeness review flags,
-and self-consistency review flags.
-
-mode:
-  - agentic — gate before the tool loop (legacy path; also used for direct agentic entry)
-  - route   — gate *before* quaternary semantic routing so localchat/webchat
-              also get executable soft outcomes (defer / clarify / degrade_chat)
+Gate logic now lives with edge cognitive state in :mod:`cognition.attention`.
 """
-from __future__ import annotations
-
-import re
-
-_CRITICAL_TASK_RE = re.compile(
-    r"\b("
-    r"urgent|emergency|asap|right now|immediately|"
-    r"safety|danger|hurt|injured|crisis|"
-    r"deadline today|due today|production (?:is )?down|outage|"
-    r"approve run-|cancel (?:the )?(?:job|schedule|reminder)"
-    r")\b",
-    re.I,
+from cognition.attention import (  # noqa: F401
+    capability_from_outcomes,
+    is_critical_task,
+    is_time_sensitive,
+    should_attempt,
+    soft_user_prompt,
 )
-_WORD_RE = re.compile(r"[a-z0-9_]{3,}")
-_TIME_SENSITIVE_RE = re.compile(
-    r"\b(?:latest|today|tonight|tomorrow|yesterday|now|current|recent|live|breaking|right now|asap|deadline|due)\b",
-    re.I,
-)
-# Modes where soft opt-out (defer/clarify/degrade) is allowed.
-_SOFT_MODES = frozenset({"agentic", "route"})
-
-
-def _tokens(text: str) -> set[str]:
-    return set(_WORD_RE.findall((text or "").lower()))
-
-
-def is_critical_task(user_input: str) -> bool:
-    return bool(_CRITICAL_TASK_RE.search(user_input or ""))
-
-
-def is_time_sensitive(user_input: str) -> bool:
-    return bool(_TIME_SENSITIVE_RE.search(user_input or ""))
-
-
-def capability_from_outcomes(outcomes: list[dict], domain: str = "") -> dict:
-    """Synthesize recent tool outcomes into a coarse domain confidence.
-    
-    dormant analysis: recent tool success/failure patterns.
-    Returns capability profile with success_rate, confidence tier, and avoid flag.
-    """
-    domain = (domain or "").strip().lower()
-    rows = list(outcomes or [])
-    if domain:
-        rows = [
-            o for o in rows
-            if domain in str(o.get("tool") or "").lower()
-            or domain in str(o.get("detail") or "").lower()
-        ]
-    n = len(rows)
-    if n == 0:
-        return {
-            "domain": domain or "any",
-            "samples": 0,
-            "success_rate": None,
-            "confidence": "unknown",
-            "avoid": False,
-        }
-    successes = sum(1 for o in rows if o.get("ok"))
-    rate = successes / n
-    avoid = n >= 3 and rate <= 0.34
-    confidence = "high" if n >= 4 else "moderate" if n >= 2 else "low"
-    return {
-        "domain": domain or "any",
-        "samples": n,
-        "success_rate": round(rate, 3),
-        "confidence": confidence,
-        "avoid": avoid,
-    }
-
-
-def soft_user_prompt(user_input: str, action: str, reason: str = "") -> str:
-    """Single structured prompt for soft gate outcomes (no multi-step tasks).
-
-    Replaces ad-hoc "[Internal note…]" blocks scattered in think.py.
-    The model must not mention system details; one short in-character reply only.
-    """
-    text = (user_input or "").strip()
-    action = (action or "").strip().lower()
-    # reason is logged/recorded elsewhere; keep user-facing prompt free of internals.
-    if action == "defer":
-        return (
-            f"{text}\n\n"
-            "[Style only — never quote this. One short in-character line: "
-            "you are low-energy and this is not urgent; offer to continue later.]"
-        )
-    if action == "clarify":
-        return (
-            f"{text}\n\n"
-            "[Style only — never quote this. Ask exactly one concrete clarifying "
-            "question about what they need. Do not start a multi-step task.]"
-        )
-    # degrade_chat / unknown → pass through unchanged
-    return text
-
-
-def should_attempt(
-    *,
-    user_input: str,
-    mode: str = "agentic",
-    energy: float = 0.5,
-    uncertainty: float = 0.0,
-    tool_outcomes: list[dict] | None = None,
-    contradictions: list[str] | None = None,
-    response_reviews: list[dict] | None = None,
-    load: float = 0.0,
-    trend_data: dict | None = None,
-    intent_confidence: float = 0.0,
-    intent_success_prob: float = 0.5,
-    enabled: bool = True,
-) -> tuple[bool, str, str]:
-    """Return (ok, reason, action) with action in proceed|degrade_chat|defer|clarify.
-
-    The gate combines the cheap readiness signal (energy/load) with bounded
-    reliability/safety signals: uncertainty, recent tool outcomes, contradictions,
-    time sensitivity, answer-completeness review flags, self-consistency review
-    flags, trend analysis, and optional intent classification confidence.
-    
-    Trend data:
-      - energy_trend: 'rising'|'falling'|'steady'|'unknown'
-      - capability_trend: 'improving'|'degrading'|'stable'|'unknown'
-      - uncertainty_trend: 'rising'|'falling'|'steady'|'unknown'
-    
-    Intent confidence (from XGBoost, if available):
-      - intent_confidence ∈ [0, 1]: how certain the model is
-      - intent_success_prob ∈ [0, 1]: predicted success rate
-    """
-    if not enabled:
-        return True, "attempt gate disabled", "proceed"
-    text = (user_input or "").strip()
-    if not text:
-        return True, "empty input", "proceed"
-    if is_critical_task(text):
-        return True, "critical request", "proceed"
-    if is_time_sensitive(text):
-        return True, "time-sensitive request should proceed with verification discipline", "proceed"
-
-    mode_norm = (mode or "").strip().lower()
-    soft = mode_norm in _SOFT_MODES
-    energy_v = max(0.0, min(1.0, float(energy or 0.0)))
-    load_v = max(0.0, min(1.0, float(load or 0.0)))
-    
-    # Option 1: Trend risk check (engaged first, before capability check)
-    trend = trend_data or {}
-    risk_signals = sum([
-        trend.get("energy_trend") == "falling",
-        trend.get("capability_trend") == "degrading",
-        trend.get("uncertainty_trend") == "rising",
-    ])
-    if soft and risk_signals >= 2 and energy_v < 0.45:
-        return False, "multiple degrading trends detected; prefer lighter path", "degrade_chat"
-    
-    # Option 2: Intent classifier check (XGBoost, if available and trained)
-    # Confidence > 0 means model was available and made a prediction.
-    intent_conf = max(0.0, min(1.0, float(intent_confidence or 0.0)))
-    intent_prob = max(0.0, min(1.0, float(intent_success_prob or 0.5)))
-    if soft and intent_conf > 0.65 and intent_prob < 0.35:
-        return False, f"intent classifier: low success probability ({intent_prob:.1%})", "clarify"
-    
-    # Existing checks follow (unchanged)
-    if mode_norm == "agentic":
-        text_tokens = _tokens(text)
-        scoped_outcomes = [
-            o for o in (tool_outcomes or [])
-            if text_tokens & _tokens(f"{o.get('tool') or ''} {o.get('detail') or ''}")
-        ]
-        cap = capability_from_outcomes(scoped_outcomes, "")
-    else:
-        # Pre-route: no reliable task/tool scope — skip avoid-rule.
-        cap = {
-            "domain": "any",
-            "samples": 0,
-            "success_rate": None,
-            "confidence": "unknown",
-            "avoid": False,
-        }
-
-    text_tokens = _tokens(text)
-    related_contradictions = [
-        c for c in (contradictions or [])
-        if len(text_tokens & _tokens(c)) >= 1
-    ]
-    latest_review = (response_reviews or [{}])[0] if response_reviews else {}
-    latest_flags = [str(f) for f in latest_review.get("flags", [])] if isinstance(latest_review, dict) else []
-    incomplete_flag = any("may not answer" in flag or "completeness" in flag for flag in latest_flags)
-    self_consistency_flag = any("self" in flag and ("conflict" in flag or "consistency" in flag) for flag in latest_flags)
-
-    if soft and load_v >= 0.75 and energy_v < 0.45:
-        return False, "running hot and energy low; discretionary work can wait", "defer"
-
-    if soft and energy_v < 0.28:
-        return False, "energy low; discretionary work can wait", "defer"
-
-    if soft and uncertainty > 0.62:
-        if len(text) < 80 and ("?" in text or len(_tokens(text)) <= 6):
-            return False, "uncertainty elevated; need a clearer ask", "clarify"
-        return False, "uncertainty elevated; prefer lighter chat path", "degrade_chat"
-
-    if soft and related_contradictions:
-        return False, "current prompt overlaps unresolved contradictions; clarify before acting", "clarify"
-
-    if soft and cap.get("avoid") and (cap.get("samples") or 0) >= 3:
-        return False, "recent tool outcomes are weak; prefer lighter path", "degrade_chat"
-
-    if soft and (incomplete_flag or self_consistency_flag) and len(text_tokens) <= 6:
-        return False, "last answer review found a likely issue; clarify the follow-up before acting", "clarify"
-
-    return True, "self-assessment clear", "proceed"
